add: UserAssignedProject
id pid uid level로 이루어진 유저 프로젝트 할당 기능 추가/ junit 테스트 통과 완

+++ src/main/java/com/example/demo/service/UserAssignService.java
@@ -2,12 +2,16 @@ package com.example.demo.service;
 
 import com.example.demo.domain.Member;
 import com.example.demo.domain.Project;
+
 import com.example.demo.domain.UserAssignProj;
 import com.example.demo.dto.UserAssignDto;
+
 import com.example.demo.repository.MemberRepository;
 import com.example.demo.repository.ProjectRepository;
 import com.example.demo.repository.UserAssignRepository;
 import lombok.RequiredArgsConstructor;
+import org.json.JSONArray;
+import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -24,15 +28,6 @@ public class UserAssignService {
     private final MemberService memberService;
     private final ProjectRepository projectRepository;
     private final MemberRepository memberRepository;
-
-    @Autowired
-    public UserAssignService(UserAssignRepository userAssignRepository, ProjectService projectService, MemberService memberService, ProjectRepository projectRepository, MemberRepository memberRepository) {
-        this.userAssignRepository = userAssignRepository;
-        this.projectService = projectService;
-        this.memberService = memberService;
-        this.projectRepository = projectRepository;
-        this.memberRepository = memberRepository;
-    }
 
     public String assignUserToProject(UserAssignDto userAssignDto) {
         String projectTitle=userAssignDto.getProjectTitle();
@@ -59,6 +54,46 @@ public class UserAssignService {
         return "true";
     }
 
+    public String getAssignment(Integer id){
+        Optional<UserAssignProj> userAssignProj=userAssignRepository.findById(id);
+        if(userAssignProj.isEmpty()){return "";}
+
+        JSONObject jsonObject = new JSONObject();
+        Optional<Member> user=memberRepository.findById(userAssignProj.get().getUid());
+        Optional<Project> project=projectRepository.findById(userAssignProj.get().getPid());
+        jsonObject.put("project", project.get().getTitle());
+        jsonObject.put("user",user.get().getFullName());
+        jsonObject.put("level",memberRepository.findById(userAssignProj.get().getUid()).get().getLevel());
+        return jsonObject.toString();
+    }
+
+    public String getAssignmentByProject(Integer pid){
+        List<UserAssignProj> userAssignProj=userAssignRepository.findAllByPid(pid);
+        if(userAssignProj.isEmpty()){return "";}
+        JSONArray assign = new JSONArray();
+        userAssignRepository.findAllByPid(pid).forEach(item ->{
+            JSONObject jsonObject = new JSONObject();
+            jsonObject.put("project", projectRepository.findById(item.getPid()).get().getTitle());
+            jsonObject.put("user",memberRepository.findById(item.getUid()).get().getFullName());
+            jsonObject.put("level",memberRepository.findById(item.getUid()).get().getLevel());
+            assign.put(jsonObject);
+        });
+        return assign.toString();
+    }
+    public String getAssignmentByMember(Integer uid){
+        List<UserAssignProj> userAssignProj=userAssignRepository.findAllByUid(uid);
+        if(userAssignProj.isEmpty()){return "";}
+        JSONArray assign = new JSONArray();
+        userAssignRepository.findAllByUid(uid).forEach(item ->{
+            JSONObject jsonObject = new JSONObject();
+            jsonObject.put("project", projectRepository.findById(item.getPid()).get().getTitle());
+            jsonObject.put("user",memberRepository.findById(item.getUid()).get().getFullName());
+            jsonObject.put("level",memberRepository.findById(item.getUid()).get().getLevel());
+            assign.put(jsonObject);
+        });
+        return assign.toString();
+    }
+
     public Optional<UserAssignDto> findAssignmentById(Integer id) {
         return userAssignRepository.findUserAssignProjById(id)
                 .map(userAssignProj -> convertToDto(userAssignProj,
@@ -80,15 +115,24 @@ public class UserAssignService {
                 .collect(Collectors.toList());
     }
 
-    public void deleteAssignmentById(Integer id) {
+//    public Integer getId(String username, String projectTitle) {
+//        Integer uid=memberRepository.findByName(username).getId();
+//        Integer pid=projectRepository.findByTitle(projectTitle).getId();
+//        Integer aid=userAssignRepository.findByUidAndPid(uid,pid).get().getId();
+//        return aid;
+//    }
+
+    public String deleteAssignmentById(Integer id) {
+        if(userAssignRepository.findById(id).isEmpty()){return "false";}
         userAssignRepository.deleteById(id);
+        return "true";
     }
 
     private UserAssignDto convertToDto(UserAssignProj userAssignProj, String projectTitle, String username) {
         return UserAssignDto.builder()
                 .projectTitle(projectTitle)
                 .username(username)
-                .level(Member.getStringFromLevel(memberRepository.findByName(username).getLevel()))
                 .build();
     }
+
 }
